# Conflicts:
#	lib/app/screens/portfolio/view/components/empty_portfolio_body/components/empty_portfolio_body_image.dart
#	lib/app/shared/models/currency_model.dart
#	lib/app/shared/providers/currencies_pod/currencies_pod.dart

+++ lib/app/shared/features/currency_buy/notifier/preview_buy_with_asset_notifier/preview_buy_with_asset_notifier.dart
@@ -1,5 +1,6 @@
 import 'dart:async';
 
+import 'package:decimal/decimal.dart';
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:logging/logging.dart';
@@ -42,7 +43,7 @@ class PreviewBuyWithAssetNotifier
 
   void _updateFrom(PreviewBuyWithAssetInput input) {
     state = state.copyWith(
-      fromAssetAmount: double.parse(input.amount),
+      fromAssetAmount: Decimal.parse(input.amount),
       fromAssetSymbol: input.fromCurrency.symbol,
       toAssetSymbol: input.toCurrency.symbol,
     );
